Rename cucumber to gherkintools
