+++ test/gherkintools/doc.go
@@ -1,4 +1,4 @@
-// Package cucumber provides tools for working with Gherkin.
+// Package gherkintools provides tools for working with Gherkin.
 // This is a separate package to allow including these tools in both test and test/steps.
 // This package is not named "Gherkin" to avoid collisions with the "gherkin" package from godog.
-package cucumber
+package gherkintools
