+++ test/gherkintools/rendering.go
@@ -1,4 +1,4 @@
-package cucumber
+package gherkintools
 
 import (
 	"fmt"
